add session check and redirections

+++ php/delete-product.php
@@ -1,11 +1,11 @@
 <?php
 session_start();
 include 'functions.php';
-if(check_user()){
+if(isset($_POST['id']) && check_user()){
+    //delete from database
     $pdo = pdo_connect_mysql();
-    // Check that the product ID exists
-    if (isset($_POST['id'])) {
-        $stmt = $pdo->prepare('DELETE FROM products WHERE id_products = ?');
-        $stmt->execute([$_POST['id']]);
-    }
+    $stmt = $pdo->prepare('DELETE FROM products WHERE id_products = ?');
+    $stmt->execute([$_POST['id']]);
+}else{
+    header('Location:../login.php');
 }
